requirements.txt
autopep8==1.7.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.14
cfgv==3.3.1
charset-normalizer==2.1.1
chatgptpy==1.0.8
colorama==0.4.6
cssselect2==0.7.0
distlib==0.3.6
filelock==3.8.0
identify==2.5.5
idna==3.4
load-dotenv==0.1.0
lxml==4.9.1
nodeenv==1.7.0
numpy==1.23.3
pandas==1.5.0
Pillow==9.2.0
platformdirs==2.5.2
pre-commit==2.20.0
protobuf==4.21.9
pyChatGPT==0.1.0
pycodestyle==2.9.1
pyperclip==1.8.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
reportlab==3.6.12
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
svglib==1.4.1
tinycss2==1.2.1
tls-client==0.1.5
toml==0.10.2
tornado==6.2
urllib3==1.26.12
uuid==1.30
virtualenv==20.16.5
webencodings==0.5.1
